feat(faq): update FAQ URLs and implement new FAQ page with SEO optimization
